oct23

+++ VTBuilder/src/main/java/com/jay9971/VTBuilder/Repository/Users.java
@@ -51,7 +51,7 @@ public class Users {
 		this.teamRating = teamRating;
 	}
 
-	public int getPlayerAccuracy() {
+	public double getPlayerAccuracy() {
 		return playerAccuracy;
 	}
